requirements.txt
streamlit
ultralytics
pillow
opencv-python-headless
streamlit-webrtc
av
requests